add user plans

+++ app/Controllers/Http/TodoController.js
@@ -4,13 +4,13 @@ const { validateAll } = use('Validator');
 const Todo = use('App/Models/Todo');
 
 class TodoController {
-  async index ({ view, auth }) {
+  async index({ view, auth }) {
     const todos = await auth.user.todos().fetch();
-    
+
     return view.render('home', { todos: todos.toJSON() });
   }
 
-  async store ({ request, response, session, auth }) {
+  async store({ request, response, session, auth }) {
     const body = request.all();
 
     const todo = await auth.user.todos().create({
@@ -21,11 +21,19 @@ class TodoController {
   }
 
 
-  async edit({ request, view }) {
+  async edit({ request, response, session, view, auth }) {
+    if (auth.user.plan === 'free') {
+      session.flash({ notification: 'You need to be on a premium plan to be able to edit.' });
+      return response.redirect('back');
+    }
     return view.render('edit-todo', { todo: request.todo });
   }
 
-  async update({ response, request, session, params }) {
+  async update({ response, request, session, params, auth }) {
+    if (auth.user.plan === 'free') {
+      session.flash({ notification: 'You need to be on a premium plan to be able to edit.' });
+      return response.redirect('back');
+    }
     const { todo } = request;
     todo.text = request.all().text;
     await todo.save();
@@ -34,7 +42,11 @@ class TodoController {
     return response.redirect('/dashboard');
   }
 
-  async destroy({ request, response, session, params }) {
+  async destroy({ request, response, session, params, auth }) {
+    if (auth.user.plan === 'free') {
+      session.flash({ notification: 'You need to be on a premium plan to be able to edit.' });
+      return response.redirect('back');
+    }
     const { todo } = request;
 
     await todo.delete();
